0.1.1

+++ src/main/java/com/develop/dental_api/repository/ServiceRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.develop.dental_api.model.entity.Service;
 
 @Repository
-public interface ServiceRepository extends JpaRepository<Service, Long> {
+public interface ServiceRepository extends JpaRepository<Service, Integer> {
 }
